+++ femlib/src/EMeshStiffness.cpp
@@ -1,21 +1,21 @@
 #include "Element.hpp"
 #include "ElementMesh.hpp"
 #include "Material.hpp"
-#include "MatrixXd.hpp"
+#include "MatrixX.hpp"
 
-MatrixXd ElementMesh::getStiffness(int eIdx)
+MatrixXf ElementMesh::getStiffness(int eIdx)
 {
-  MatrixXd K = m[me[eIdx]]->getStiffness(e[eIdx],this);
+  MatrixXf K = m[me[eIdx]]->getStiffness(e[eIdx],this);
   return K;
 }
 
-MatrixXd ElementMesh::getStiffness()
+MatrixXf ElementMesh::getStiffness()
 {
   int matSize = 3 * (int)x.size();
-  MatrixXd Kglobal(matSize,matSize);
+  MatrixXf Kglobal(matSize,matSize);
   Kglobal.fill(0);
   for(unsigned int ii = 0;ii<e.size();ii++){
-    MatrixXd K = getStiffness(ii);
+    MatrixXf K = getStiffness(ii);
     for(int jj = 0; jj<e[ii]->nV(); jj++){
       int vj = e[ii]->at(jj);
       for(int kk = 0; kk<e[ii]->nV(); kk++){
